# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个基于 Go 语言构建的轻量级、高性能通用电商后端系统。",
        "title": "GoShop API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3233",
    "basePath": "/",
    "paths": {
...
        "/api/products": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "分页查询商品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类 ID",
                        "name": "categoryId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "查询商品详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SPU ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Spu"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Sku": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "description": "单位：分",
                    "type": "integer"
                },
                "salesVolume": {
                    "type": "integer"
                },
                "specs": {
                    "description": "存储规格JSON属性",
                    "type": "string"
                },
                "spuId": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Spu": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "detailHtml": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "description": "存储图片JSON数组",
                    "type": "string"
                },
                "mainImage": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "skus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Sku"
                    }
                },
                "status": {
                    "description": "1: 上架, 2: 下架",
                    "type": "integer"
                },
                "subtitle": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}